services/user-service/requirements.txt
fastapi
uvicorn
databases[postgresql]
sqlalchemy
passlib[bcrypt]
pyjwt
pydantic[email]
psycopg2
python-dotenv